Removed static from common actions
