+++ DFC.App.JobProfiles.HowToBecome.Tests.API.IntegrationTests/DFC.App.JobProfiles.HowToBecome.Tests.API.IntegrationTests.csproj
@@ -31,4 +31,8 @@
     </None>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Test\JSON\" />
+  </ItemGroup>
+
 </Project>
